requirements.txt
requests
faster-whisper
webrtcvad-wheels
torch
torchaudio
torchvision
customtkinter
numpy
pyaudio
ollama
pydub